requirements.txt
Flask
Flask-SQLAlchemy
IPython
Pillow